fix mandatory listening on default TCP endpoint
The dotsd currently always listens on the default localhost TCP
endpoint. This makes it impossible to set up dotsd instances that don't
use TCP at all or want to use the default port for something else.

Instead, the default endpoint must only be added when the user did not
specify any endpoint themselves.