disallow mixing of legacy and endpoint dotsd args
It is currently possible to mix endpoint and legacy arguments in the
dotsd, resulting in unintuitive behavior.

For example, even if a port is set within a listen TCP endpoint
argument, it can get overridden by a legacy argument.

Instead, the legacy arguments should only be used when the listen
endpoint arguments are defaulted (i.e. unspecified), so that the two
methods cannot be mixed.