introduce overhauled expectation sequences
The library currently already supports the expectation of sequences via
call expectations.

However, (consecutive) call expectations are flawed in multiple ways.
Among other things, they are lacking in the following ways:

 - They are unable to include negative default expectations (e.g.
 expecting that a mock function is never called).

 - They produce hard to dissect error messages, because the default
 expectations are technically created from the same source location.

 - They can be unintuitive to read, because default expectations are
 created implicitly from arguments (e.g. DOTS struct types).

 - They are technically complex, demanding to compile and take a lot of
 convoluted macro work to extend and maintain.

As a replacement, this commit introduces support for generalized
expectation sequences.

Expectation sequences combine the advantages of both consecutive and
named call sequences, but do not use a default expectation factory.
Instead, all expectations always have to be created directly or via a
factory argument.

This results in flatter, simpler and easier to extend structures, with
the option to retain full source location information.

Note that the existing call expectation support is not removed to
retain backwards compatibility.