migrate publish/transmit support to new sequences
This migrates the support for publish and transmit expectations,
including the corresponding base classes, to the recently introduced
expectation sequences (see [1]).

The migration comes with an overhaul of the test interface. To increase
readability, all expectations now have to be specified explicitly. This
results in the PublishExpectation and TransmitExpectation being
replaced by corresponding macros. For consistency, this change is also
reflected in spoofing publishes and transmits.

Additionally, this also changes the naming scheme for the test support
macros, with the goal to increase clarity in sequences with multiple
types of expectations.

Expectation macros now always start with 'EXPECT_' (analogous to the
GTest EXPECT_CALL macro), while all spoof macros start with 'SPOOF_'.

Note that this also deprecates all call expectation support, which has
to explicitly be re-enabled by a corresponding preprocessor definition.

References:

 - [1] commit 7f788119151ce415d7166c925bdb119bb5ddb24a