support publish expectations on all transceivers
Publish expectations can currently only be created in association with
a guest transceiver. Additionally, those expectations are always
created on the host.

Not only does this prevent creating (clear) expectations associated
with the host itself, it also does not allow expectations on guests
themselves (e.g. for self publishes).

To rectify this, expectation creation can be overhauled to be
associated with transceivers, instead of guests on the host.

Note that this results in now having to manually add non-default guest
connections.