use events for publish expectations
Publish expectations were originally based on transmissions, because
the assumption was that events were not needed and a solution for
events more fiddly to implement.

However, transmissions have the disadvantage that their handlers are
invoked before the container is updated. This means that, for example,
corresponding actions for self publish expectations cannot rely on the
container content, even though that is exactly what the expectation is
about.

Additionally, events are more flexible and intuitive to use in
corresponding actions, because most users know them from regular
subscription handlers.

Note that this change also removes the delay for the invoking of mock
subscription handlers (see [1]). The delay is no longer needed, because
a user can now explicitly set an expectation for the self update and
use the corresponding action instead.

The delay also caused some issues in the past, because it creates an
unintuitive order of events in the IO context.

References:

 - [1] commit 9e9a06e489e1b4b55c87be5e0a3e316b84a0806c