remove factory support from expectation sequences
Expectation sequences currently support expectation factories as
arguments in addition to expectations themselves. This is a leftover
from the call sequence support on which expectation sequences were
based on.

However, since the factories are invoked immediately when the
expectation sequence is created, there is no point in supporting them
because they provide no benefit to just passing the expectation
directly.