support partial ordering in expectation sequences
Arguments of expectation sequences are currently always totally
ordered.

However, there are many use cases in which the relative order of some
expectations within a sequences are either unknown or should not be
included as part of the test.

Examples for this are situations in which the order of events is
determined only by implementation details, such as whether an std::set
or an std::unordered_set is used to manage data.

Instead, a user should be able to specify sets of expectations to
define partial orderings. In practice this means that all expectations
in a set can occur in any order, while the whole set is treated like a
single expectation within the outer sequence.

Note that the corresponding class for typed expectation sets introduced
with this is currently only intended to be used internally.

Also note that expectation sequences no longer rely on Google Test
sequences and instead use the After clause of expectations. As such,
the possibility to use an externally created sequence has to be
removed.