rename publish test base to event test base
Even though publish expectations are a core part of the testing
support, the base class can also be used to test components that only
rely on subscriptions; i.e. DOTS events.

It makes therefore more sense to refer to the overarching concept of
events, instead of just publishing, which a component might not perform
itself.

Additionally, this avoids the name clash with the deprecated
PublishTestBase class and eases migration to the new components, by
allowing both to be used simultaneously.