handle cardinalities in expectation sets
Expectation sets currently always assume a cardinality of 1 when adding
an action.

This is neither safeguarded nor documented and can result in unexpected
behavior with unintuitive errors.

Instead, expectation sets should properly handle cardinalities. This
includes proper error handling, as well as supporting members with a
cardinality of 0.

Note that Google Test offers only limited access to cardinality
information, which is why actions are generally not supported on
expectations with cardinalities greater than 1.