fix redundant CMake find files for RapidJSON
In the process of adding CMake rules for packaging, a redundant CMake
file for finding RapidJSON was added unintentionally (see [1]). This
means that the repository currently holds two such files that only
differ in capitalization of their names.

On file systems that are case insensitive (e.g. NTFS on Windows), this
causes issues, because only one of the files will be checked out. If
the same repository is then used in a case sensitive context (e.g. by
mounting the directory into a Linux Docker container), builds might
fail, because the filenames CMake is looking for are determined by the
exact package name (i.e. find_package(RapidJSON) vs.
find_package(Rapidjson)).

Since RapidJSON is the correct package name, FindRapidjson.cmake should
be removed to resolve this issue.

References:

 - [1] commit e3d58c4d070b00c316904863da5a9947fcae4402