use consistent parameter style for std::function
Handlers based on std::function are currently passed inconsistently: by
value, by const-ref and by rvalue-ref.

Even though it usually makes only little difference from a performance
point of view, the interfaces should use a consistent style for
semantical reasons.

Passing a handler by constant reference implies to a certain degree
that the argument is not copied. This can make a difference, if, for
example, a mutable lambda is used and the user assumes a shared state
between all instances where the handler is given.

Using an rvalue reference implies that a user must give up ownership of
the handler. Although this prevents unintentionally passing an lvalue
reference under the assumption of shared state, it also restricts
reusing a handler object multiple times, without manually creating
copies.

On the other hand, even though passing a handler by value does not
technically prevent the shared state pitfall, it makes it clear that it
will be copied, while also retaining flexibility in passing the
argument. It is also the simplest and most readable variant.

For above reasons, handler parameters should be refactored as follows:

 - Use pass-by-value if the function creates a copy to invoke the
 handler after the function returns.

 - Use pass-by-const-ref if the handler is invoked before the function
 returns, to avoid unnecessary copies and allow repeated usage of the
 handler object.