fix value type access for properties named 'value'
When a property is named 'value', the code generator will currently
generate a property type with the name 'value_t'.

If a component attempts to access the value type of that property by
using 'typename P::value_t', the build will fail, because the name
'value_t' technically refers to the constructor 'value_t::value_t' and
not the value type alias of the Property base class.

The issue for example currently occurs when a user is attempting to use
a type visitor for a corresponding DOTS struct type (e.g. by calling
dots::to_string()).

A similar issue was already fixed for the property initializers some
time ago (see changes including a corresponding code comment in [1]).

To resolve the problem entirely, the template for DOTS struct types
should be changed to just use a suffix other than '_t' to designate
property types.

References:

 - [1] commit 385cec181c387a7b8e41630cbee8fcc4e4ad338d