simplify publishing of partial instances
When publishing a subset of valid properties, the given set currently
must always include the key properties.

However, this requirement is redundant, because the key properties
always have to be valid by definition anyway.

Instead, a user should be able to specify a set of non-key properties,
which then will be merged with the key property set automatically.