safeguard validity of key properties on receive
Connections currently do not check whether all key properties of a
received instance are valid.

This can result in unexpected behavior when instances with missing key
properties are dispatched both by a guest or host.

Because connections do not know whether a peer implements the DOTS
protocol correctly, the validity of key properties always has to be
checked when a transmissions is received.