optimize property iteration performance
The property iterator currently always emplaces a proxy property when
iterating over the given property set, even if the property is skipped.

This means that iteration always incurs the costs of creating the
proxy, even though it would be sufficient to just check the property
set via the property descriptor.

This optimization can improve performance significantly for operations
based on property iteration, such as Struct::_less.