fix transparent key compare of container
The key compare functor of the container is currently incomplete. This
can result in unintended constructions of AnyStruct objects (and
therefore copies) when operations on the underlying container map are
performed with a non-AnyStruct argument (e.g. const type::Struct&).

This can significantly impact performance in (very) high load
scenarios, when thousands of cached instances are inserted and removed
and container performance is a bottleneck.

Note that the issue does not seem to appear with all compilers. For
example: It was observed with GCC 9.3, but not MSVC.