optimize lookups of shared types in channels
Using the emplace function of an std::set to check whether a type is
shared may result in construction of the string element even if it is
already part of the set (see [1]).

Since copying the type name string on every transmit and receive can
significantly impact performance, the lookups and insertions should be
performed in separate operations.

References:

 - [1] https://en.cppreference.com/w/cpp/container/set/emplace