overhaul and optimize stream based channels
Channels that are using Asio streams (i.e. TCP and UNIX domain socket
channels) are currently based on a simple "one at a time" approach.

While straightforward in implementation, this has several severe
drawbacks in terms of scalability: Writing synchronously means that a
host application has to wait for an object to be written for every
connection that is subscribed to that type. While reads are performed
asynchronously, they consist of multiple operations per object and only
ever read one object at a time.

To increase CPU utilization and keep the number of system calls low,
stream channels can be overhauled to operate asynchronously only and
use buffering in both directions; which can then be used to read and
write multiple payloads at once.

Additionally, the transmission format can be optimized and payloads be
cached to avoid redundant serializations for DOTS hosts.

Overall, these changes can dramatically improve throughput performance,
up to several orders of magnitude in extreme cases.