optimize exporting of dependencies in channels
Since the lookup to check whether a type needs to be exported is
performed based on an explicit descriptor, using the type name is
unnecessarily costly.

Instead, the check can exploit the uniqueness of a descriptor's pointer
value, which can efficiently be stored and looked up in an unordered
map.