fix premature expiration of local channels
A local channel currently does not deliver its transmission when it is
itself expired.

However, after completion of the post, there is no lifetime dependency
anymore to the transmitting peer. This can result in transmits getting
canceled unnecessarily.

This issue is especially relevant when a connection is attempted to be
closed gracefully, which currently does not always work, even if the
receiving channel still does exist.