optimize updating of container instances
The container currently uses the copy operation of structs to update an
instance on insertion or removal.

However, as with the key compare, the container can make certain
assumptions to optimize the updating procedure and also avoid the
overhead of using proxy properties.

Additionally, having the implementation visible usually offers more
optimization options to the compiler.

Even though the difference is not dramatic, it is still significant,
particularly for host transceivers, where updating the container is a
considerable part of the workload.