limit size of async write buffer
The async write buffer can currently be filled indefinitely until the
previous async write is finished.

However, when a peer is not able to read any data, the buffer may grow
until the system is out of memory.

To avoid this, the write buffer should have an upper limit, so that
connections of unresponsive peers are closed at some point.

The limit should not be set too low to allow efficient transmits of
larger bursts and rather be seen as a fail-safe for extreme and
unlikely situations.

Note that the initial limit set by this commit is 10 MiB, which is
sufficient for a combined payload of roughly 100'000 medium sized
instances.