introduce universal handlers
The subscribe functions currently manually wrap arguments to support a
variety of user handlers, including member functions.

Because this functionality is already implemented redundantly and is
cumbersome to replicate elsewhere, it should be moved to a dedicated
handler class.

Note that the new class behaves similarly to std::function, with the
exception that objects always contain a valid value.