use universal handlers for typed event subscribes
Because universal handlers now provide the same functionality, the
typed event subscribe functions can be simplified by using handlers as
arguments.

This step is also necessary to provide a consistent API for using
handlers within the library.

Note that the current implementation will be retained as deprecated for
backwards compatibility.