fix usage of preceding header data in legacy reads
When using the legacy transmission format, the transport header
instance is currently not reset.

Because of this, property values of previous reads can be present in a
subsequent read when they are not overwritten by the new transmission.

This can cause incorrect behavior; e.g when a transmission is falsely
using the remove flag of a preceding one.