resolve confusing naming of text based serializers
Serializers based on text formats are currently categorized as "string
serializers" and use a corresponding base and options class.

This is confusing, because the term "string" refers to both a
particular serializer, as well as a serializer category.

To remedy this, the term "text serializer" can be used to refer to all
serializers that use a format based on human readable text.