make text specific serializer tests more readable
The test cases that are specific to text based serializers are
currently generically implemented in a corresponding test base class to
reduce the amount of duplicated test code.

However, because the text serializers can differ significantly in terms
of permitted input and expected output, this only results in a small
benefit at the cost of hard to read and maintain tests.

Because of this, the tests should rather be implemented explicitly.