simplify confusing text output style options
There are currently four different style options for the text based
serializers and it is often not obvious what the difference between
them is.

In particular, it is not clear how the 'Compact' option differs from
the 'Minimal' and 'SingleLine' ones.

To rectify this, the options can be simplified as follows:

 - Remove the current 'SingleLine' option entirely, since it is almost
 never used in practice anyway.

 - Rename the current 'Compact' option to 'SingleLine' and use it as a
 default.

 - Rename the current 'Minimal' option to 'Minified' with the intention
 that a more familiar term conveys more clearly for what the option
 should be used.