always use plain text format for top-level values
The text readers and writers, and subsequently the text serializers,
currently treat top-level values of fundamental types as if they were
part of an object or an array.

This usually leads to unintuitive formatting, when, for example,
serialization of a time point yields a quoted string and also requires
one as an input.

This is often not desirable, because top-level text conversion is
usually not done in the context of serialization, but rather for
printing log output and parsing user input (e.g. via dots::to_string()
and dots::from_string()).

Instead, all formatting rules should only be applied for values inside
of objects or arrays, while top-level values always use the plain
version.

This results in the following formatting rules for top-level values:

 - Strings are not delimited and are not escaped (this also includes
 time points and UUIDs)

 - Enums do not have type prefixes

 - Integers and floats do not have suffixes

Note that in the last two cases, the prefixes and suffixes are still
permitted as an input, albeit not mandatory.