permit serializers to discard unknown properties
Both the CBOR and the text based serializers currently expect struct
instances to only have those properties that are defined in the type
descriptor.

However, there are various circumstances in which a serializer might
encounter instances whose types do not match the descriptor, but which
are still compatible.

Since in those cases the instances are still deserializable, unknown
properties should not produce an error, but rather just be silently
handled and discarded by the serializer.