fix init value for Boost_USE_STATIC_LIBS override
The "option" command requires mandatory arguments for the name and
description, as well as an optional argument for the initial value.

Since the Boost_USE_STATIC_LIBS override option is missing the
description, the argument intended as the initial value is interpreted
as the description, while the initial value defaults to "OFF" when the
option is not set.