use reader/writer pattern for RapidJSON serializer
Even though the RapidJSON serializer is implemented separately, it
should still have an interface that is at least similar to the other
serializers.

Using the reader/writer approach also has similar benefits as in the
other implementations, such as reducing complexity and making it easier
to create alternative formats.

In addition to just moving the read and write implementation, these
changes also include additional functionality that is already present
in the other serializers:

 - The tuple interface is removed entirely, as its use case is rare and
 can be covered by using readers and writers directly.

 - The Traits type is replaced by a Format type, which uses enums
 instead of boolean switches.

 - Unknown properties are skipped automatically, allowing
 deserialization of different but compatible struct types.