consistently use newlines at the end of files
Even though this is not an issue with most modern applications, files
that are missing newlines are strictly speaking not text files, which
can have certain effects on some utilities (see [1]).

Additionally, adding to a file without a trailing newline creates
unclean diffs in Git, as the last line always has to be changed too to
add the newline character.

References:

 - [1] https://unix.stackexchange.com/questions/18743/whats-the-point-in-adding-a-new-line-to-the-end-of-a-file