consistently apply 'rule of zero'
Most classes currently explicitly define their destructor, copy and
move constructors, as well as copy and move assignment operators, even
in cases where all of them are "defaulted".

This was done with the intention to always make it explicit, which
constructors and assignment operators are going to be used in a class.

It also helps to detect, when a defaulted constructor or operator is
implicitly deleted.

However, in most cases this style results in a lot of redundant code
and reduces readability. Furthermore, using the 'rule of zero' (see
[1]), is also recommended in the C++ Core Guidelines (see [2]).

Note that while not part of the 'rule of zero', this commit also
removes explicitly defaulted default constructors when they are the
only constructor in a class.

References:

 - [1] https://en.cppreference.com/w/cpp/language/rule_of_three

 - [2] https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#cdefop-default-operations