avoid boost namespace when using Asio
Because Asio is currently used as part of Boost, accessing Asio always
has to be qualified according.

To improve readability, having to specify the boost namespace can be
avoided by using a convenience header with a namespace alias.

This also has the additional advantage that Asio can be be switched
more easily to the standalone version if necessary.