always alias fundamental types in dots namespace
The fundamental types currently reside in the dots::types namespace,
but by default are also aliased in the dots namespace.

However, because the aliasing can be disabled by a preprocessor
definition, the library has to explicitly qualify the types namespace
in many places.

Additionally, this feature is never used, because, contrary to user
types, there cannot be any name clashes with the fundamental types, due
to user defined type names.

Instead, fundamental types should always be aliased to improve
readability.