fix retirement of expectation sequence arguments
Because an argument of an expectation sequence is currently not retired
when it is satisfied, it will still be active when another matching
call occurs.

While this is not an issue with a sequence of non-matching calls, a
repeated call will fail with an "unexpected call" error, even if
another unsatisfied expectation follows in the sequence.

An example for this are two equal expectations in a row.

To fix this, the expectations have to always be retired immediately
once they are satisfied. This also has the added benefit of producing
more concise output when an expectation fails.