use generated types for static struct tests
The tests for static struct types deliberately have not used generated
types in the past to not be dependent on the generator templates.

However, because in practice static types are only used with generated
types, it does not make sense to test them separately, which only
results in requiring additional tests to verify correctness of the
templates.

Instead, the tests should directly use generated types for simplicity.