support transition handlers for all transceivers
While the host transceiver supports using a transition handler to be
notified when a connection changes state, similar support is missing
from the guest transceiver.

Among other things, this makes it currently impossible to know, when
the connection of a guest transceiver gets closed.

In general, transition handlers should be supported for all
transceivers.

Note that contrary to the current transition handler for the host
transceiver, the generalized version also passes possible errors to the
user.