resolve potential name clashes in application args
The transceiver and endpoint related command line arguments of the
dotsd and the DOTS guest application currently have generic names and
also define corresponding short options.

Since those arguments are supposed to be usable in conjunction with
user defined arguments, names like 'open' and 'listen' are not
suitable.

To reduce the likelihood of name clashes, the arguments can be prefixed
and the short options removed entirely.