directly allow opening and listening via endpoints
The creation of channels and listeners from endpoints is currently only
implemented in the Application and dotsd main.

However, there are various scenarios in which this feature is required,
but using the Application class is not possible.

This usually results in code duplication and additional maintenance
work when new channel types are added to the library.

To properly support this, opening and listening via endpoints can
directly be supported by the corresponding transceivers.