allow carefree destruction of transceivers
When a transceiver is destroyed, open connections are currently not
closed via the transition procedure.

Among other things, this results in no transition handlers being called
and the user being forced to handle that case manually.

Instead, connections should always be closed the same way, regardless
of whether this happens due to an error or destruction of the object.

Note that this also removes the recently introduced manual closing of
connections and guest transceivers, as it is no longer needed.

Furthermore, the feature is implemented poorly because it leaves the
guest transceiver in an invalid state.