allow carefree destruction of applications
When an application is destroyed, it currently does not properly clean
up its managed transceivers.

In particular, this will often prevent a graceful close of connections
when using the global guest transceiver.

Instead, destruction of the application should be carefree and ensure
that all resources are properly cleaned up and all pending async
operations are executed.