allow write completion when channels are destroyed
When an async stream channel with a non-empty write buffer is
destroyed, the write can currently not continue when the async write
handler is invoked.

Because of this, extra care is required to ensure that data is actually
written before a connection is closed.

Among other things, this makes the graceful close mechanism not
reliable in certain situations.

To resolve this, the lifetime of the channel can be extended when it is
about to be destroyed while there is still an async write operation
pending.

As long as the execution context of the stream is not destroyed as
well, this allows users to process queued writes after a connection was
closed.