allow direct access to global transceiver storage
The global guest transceiver is currently implicitly constructed when
calling dots::transceiver(), which might not be expected when using a
getter function.

Additionally, the transceiver can currently not be destroyed without
setting a new transceiver instance and proper support would require yet
another function in the global API.

Since it is rare that the global transceiver has to be accessed
directly and it is usually used in conjunction with a dots::Application
anyway, the API should be simplified by making the storage of the
global transceiver accessible directly.