prevent unintentional access of global transceiver
The name of the global transceiver getter function is currently rather
generic and often clashes with other more local functions (e.g. the one
from dots::Application).

To make it clear that a global variable is accessed, it should be
reflected in the name; similar to dots::io::global_io_context().