merge stream based channels and listeners
The default and legacy versions of both the TCP channels and listeners
are currently implemented in dedicated files, even though they are
almost identical.

To simplify this, both versions can be merged into a single channel and
listener template.

Note that this step also introduces a legacy version for the UNIX
domain socket channel and listener akin to the TCP variants.