introduce versioning of stream based channels
The stream based channels (i.e. TCP and UDS) currently only have a
default and a legacy version, which results in various issues regarding
backwards compatibility.

Because the default version currently cannot be selected explicitly,
adding another default version in the future would always result in a
breaking change.

At the same time, the term 'legacy' is ambiguous, as the referred
transmission format was also used for some time this library.

Since it is likely that more formats are added in the future and that a
certain amount of backwards compatibility is desirable, the library
should provide at least a basic form of versioning for the transmission
format, which allows users to either use the default or to select a
specific version of a channel explicitly.

This has to be possible both via endpoints, as well as from within the
library when using a channel type.