simplify comparison operations in type descriptors
It is currently mandatory for a type descriptor to define all four
comparison operations for a particular type, even though in practice
they are always implemented by defining them on the basis of the less
operation.

Since there does not seem to be a need to ever define these operations
in any other way, only the less operation should be mandatory and the
other operations implemented non-virtually in the descriptor itself.