remove filesystem support for endpoints
Endpoints currently support usage of path objects from both the STL and
boost versions of 'filesystem'.

However, this support is limited and does not provide any benefit other
than converting paths to a string.

Since it is trivial to do this manually if needed and considering the
support is barely used, it should be removed to avoid the filesystem
dependency.

In case of boost, endpoints are actually the last component that
requires boost::filesystem, so that dependency can be dropped
completely.

Note that the dependency to boost::iostreams is dropped as well, which
is also not required anymore for some time now.