fix application hang on connect error
When an error occurs while an application is waiting for a connection
to be established, the condition for leaving the loop is never met and
the control flow gets stuck.

Since the Application's constructor is running the IO context up to a
certain point before returning, any internal error should result in an
exception to produce a consistent behavior with other connection
issues, such as bad TCP connections.