use CMake build type to handle optimization levels
The optimization levels are currently fixed to O2 for the GCC and Clang
compilers and can only be changed by setting the CMAKE_CXX_FLAGS.

However, because the exact flags required to serve a certain use case
are often dependent on the platform and available compiler version,
CMake offers the CMAKE_BUILD_TYPE variable, which, when set,
automatically sets sensible flags depending on the configuration and
target system.

This variable is also used by most modern C++ IDEs and should generally
be preferred to setting the CMAKE_CXX_FLAGS directly, especially since
that variable sometimes propagates in an unexpected way through a
project structure.

For above reasons, this commit removes the fixed O2 flags and sets the
build type to 'Release' if none is specified. The code to set the
default is a slightly modified excerpt from a Kitware blog post (see
[1]).

Note that this might result in an optimization level of O3 instead of
O2 in some cases.

Also note that even after these changes the user can still not set a
build type explicitly and override the flags if desired.

References:

 - [1] https://www.kitware.com/cmake-and-the-default-build-type/