add managed/RAII timer support to global API
Timers added via the global API currently have to be managed manually
via an id.

Among other things, this makes it mandatory to define a custom
destructor when a timer is needed within a class.

Instead, the global API should return the Timer object itself, to allow
users to manage it via RAII, similar to subscriptions.

Note that the previous API, including the obsolete TimerService, is
being deprecated and will be removed in a future release.