overhaul observations of file descriptor events
Similar to timers before the recent changes, file descriptor
observations can currently only be managed in a non-RAII way, by either
using the functions from the global API or accessing the service and
FdHandler directly.

Instead of adjusting the FdHandler, it can be superseded by a
modernized RAII version, which not only refrains from using a
deprecated Asio interface, but also provides additional functionality,
such as error handling and waiting for "write readiness".

Contrary to the timer support, the current components and functions
will not be deprecated but instead be removed entirely. This is because
the FdHandler has some drawbacks, as it cannot be removed properly,
which in conjunction with the missing error handling can result in
infinite loops in certain scenarios.

Note that the FdObserver does deliberately not receive a corresponding
function in the global API, as observing file descriptors is an
advanced feature. This step also opens the door for for making the
dots.h free of Asio in the future.