make IO context arg mandatory for transceivers
The transceivers currently default to the global IO context if none is
specified.

This poses a potential hazard when an application relies on specific IO
contexts, but the author unintentionally omits the corresponding
argument.

Since the manual construction of transceivers is an advanced feature
and usually does not have to be expressed often, the default argument
should be removed for clarity and safety.