remove Asio header dependency from global DOTS API
The global API currently always has to include asio.hpp in one form or
another.

Since that header is rather big and the dots.h is usually included in
many translation units of a standard application, this introduces a
significant burden to compilers and static analysis tools.

Because Asio is primarily needed for only a few Asio components (mainly
the IO context), the dependency can be avoided by using forward
declarations instead.