allow property comparisons with left-hand values
Because the comparison operators are currently defined as members, a
property can currently not be compared to a value on the left-hand side
of the expression.

This can be resolved, by moving the operators out of the class's scope.

Note that this also moves the implementation of the property comparison
operators into the operators. This is in preparation for C++20, where
these operations will be synthesized by the three-way-comparison
operator.