fix overload resolution for writing CBOR floats
The experimental CBOR serializer does currently no longer build with
the most recent MSVC version, due to an overload resolution issue.

The compiler is trying to instantiate one of the array overloads of the
write function, which fails with the error message, that the
FloatFormat enum value is not a valid size type.

This is most likely due to the float type being specified explicitly,
causing SFINAE to not work as intended.

To fix this, the float type must no longer be specified explicitly, but
instead be deduced from the given argument.