clean up CRTP functions of properties
The property base class currently provides defaults for the CRTP
functions, that can be "overridden" by the derived property classes.

This makes it difficult to understand which functions for each property
type are actually called.

Since CRTP patterns are already not that easy to maintain, the defaults
should be removed and the required functions implemented consistently
among property types.