allow local timezone to be overridden
Time points currently always retrieve the current time zone when the
UTC option is not used.

However, there are situations in which it might be desirable to use a
different time zone from the default one.

Additionally, retrieving the current time zone can be comparatively
expensive and the user should have the option to set the time zone once
if an automatic update of the time zone is not required.

Note that this is currently only part of the experimental API.