fix list init with non value type arguments
When using the experimental designated initializers, vector properties
can currently be list-initialized via an std::initializer_list
parameter.

However, because the corresponding property constructor currently has
to be a template to hide it for non-vector properties, it is only
selected when all arguments are of the vector's value type.

Otherwise initialization will use the "take-all" constructor template
instead, which can result in either compiler errors or the usage of
non-list constructors.

This issue can be fixed by moving the list-init constructor into the
generated code, where it can be created on demand only for vector
properties, without the need for a template.