simplify static descriptor handling in property
To allow the usage of static operations when the type is known at
compile time, properties currently always have to handle the special
case of being typeless or not.

Since the static type operations are now decoupled from the specific
descriptor, the operations can also be moved directly into the
descriptor base class.

Doing so results in the static operations always be used when the type
is known at compile time, even if the descriptor is typeless itself;
making handling the special case obsolete.