fix construction/assignment from other properties
Properties are supposed to be constructable and assignable from other
properties.

However, due to an unfortunate combination of several issues, this is
currently not working as intended.

Because the is_property trait does not correctly detect properties,
corresponding arguments are passed onto the corresponding non-property
construct and assign overloads, which right now only works due to the
value type conversion operator and when the property is valid.

This issue can be resolved by fixing the trait bug and properly
handling property arguments, including allowing construction and
assignment from invalid properties.

Note that this also now allows properties to fully interact with
compatible properties of other derived types, which was previously only
possible under certain circumstances.